Extract runfaster candidate lookup ownership
